Move saving from report backend to storage for es
